+++ app/(chat)/chat/_components/chat-bubble.tsx
