+++ middleware.ts
@@ -2,7 +2,7 @@ export { default } from 'next-auth/middleware';
 
 export const config = {
   matcher: [
-    // '/chat/:path*',
+    '/chat/:path*',
     '/courses/:path*',
     '/dashboard/:path*',
     '/leaderboard/:path*',
